vae_karman/requirements.txt
torch~=2.3.1
numpy~=1.26.4
scipy~=1.13.1
pandas~=2.2.2
matplotlib~=3.9.0
tqdm~=4.66.4
torchvision~=0.18.1
pillow~=10.3.0
opencv-python~=4.10.0.82
ema-pytorch~=0.4.8
requests~=2.32.3